+++ src/oregonrealm/BufferedSocket.cpp
@@ -150,7 +150,7 @@ bool BufferedSocket::send(const char *buf, size_t len)
 
         if (n < 0)
             return false;
-        else if (n == len)
+        else if (size_t(n) == len)
             return true;
 
         // adjust how much bytes we sent
@@ -196,7 +196,7 @@ bool BufferedSocket::send(const char *buf, size_t len)
         mb->release();
         return -1;
     }
-    else if (n == mb->length())
+    else if (size_t(n) == mb->length())
     {
         mb->release();
         return 1;
@@ -245,7 +245,7 @@ bool BufferedSocket::send(const char *buf, size_t len)
     return n == space ? 1 : 0;
 }
 
-/*virtual*/ int BufferedSocket::handle_close(ACE_HANDLE h, ACE_Reactor_Mask m)
+/*virtual*/ int BufferedSocket::handle_close(ACE_HANDLE, ACE_Reactor_Mask)
 {
     this->OnClose();
